// src/plantify_admin/src/declarations/plantify_backend/plantify_backend.did.d.ts
export interface Founder {
  'id' : string,
  'linkedIn' : string,
  'principal' : Principal,
  'createdAt' : Time,
  'fullName' : string,
  'email' : string,
  'experience' : string,
  'updatedAt' : Time,
  'idNumber' : string,
  'expertise' : string,
  'taxNumber' : string,
  'address' : string,
  'phone' : string,
  'previousBusinesses' : string,
}
// ...
export type Time = bigint;